app/agentcore/requirements.txt
# AgentCore Runtime 用依存関係
# これらの依存関係は AgentCore デプロイ時に Docker イメージに含まれます

# AI エージェント基盤
strands-agents>=1.0.0
bedrock-agentcore>=1.0.0

# Python パッケージ管理（MCP Server 起動用）
uv>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# データモデル・バリデーション
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP クライアント
httpx>=0.25.0

# ログ・監視
aws-lambda-powertools>=2.0.0

# ユーティリティ
python-dateutil>=2.8.0